changed downscale reference plane
Pierre explained that downscaling should be done for the plane that
has been filtered to the equivalent of a 2*sigma filter, not the
last level of an octave. He pointed out that this is actually explained
in the original SIFT paper. The previous code was wrong, now the
reference is the last-2 level of an octive. This is always correct
if the number of levels is correct for the chosen sigma. However,
our implementation is currently correct on for the hard-coded sigma
of 1.0